Backend/pyproject.toml
[project]
name = "pse"
version = "0.1.0"
description = "Personal Semantic Search Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "chromadb",
    "sentence-transformers",
    "pypdf",
    "google-generativeai",
    "python-dotenv",
    "pydantic",
    "aiofiles",
    "setuptools",
]
